// src/protos/bilibili.community.service.dm.v1/DM.ts
export interface BuzzwordConfig {
    /**
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.BuzzwordShowConfig keywords = 1;
     */
    keywords: BuzzwordShowConfig[];
}
// ...
export interface DanmakuFlagConfig {
    /**
     * 云屏蔽等级
     *
     * @generated from protobuf field: int32 rec_flag = 1;
     */
    recFlag: number;
    /**
     * 云屏蔽文案
     *
     * @generated from protobuf field: string rec_text = 2;
     */
    recText: string;
    /**
     * 云屏蔽开关
     *
     * @generated from protobuf field: int32 rec_switch = 3;
     */
    recSwitch: number;
}
// ...
export interface DanmuPlayerViewConfig {
    /**
     * 弹幕默认配置
     *
     * @generated from protobuf field: bilibili.community.service.dm.v1.DanmuDefaultPlayerConfig danmuku_default_player_config = 1;
     */
    danmukuDefaultPlayerConfig?: DanmuDefaultPlayerConfig;
    /**
     * 弹幕用户配置
     *
     * @generated from protobuf field: bilibili.community.service.dm.v1.DanmuPlayerConfig danmuku_player_config = 2;
     */
    danmukuPlayerConfig?: DanmuPlayerConfig;
    /**
     * 弹幕显示区域自动配置列表
     *
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.DanmuPlayerDynamicConfig danmuku_player_dynamic_config = 3;
     */
    danmukuPlayerDynamicConfig: DanmuPlayerDynamicConfig[];
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.DanmuPlayerConfigPanel danmuku_player_config_panel = 4;
     */
    danmukuPlayerConfigPanel?: DanmuPlayerConfigPanel;
}
// ...
export interface DmViewReply {
    /**
     * 是否已关闭弹幕
     * 0:未关闭 1:已关闭
     *
     * @generated from protobuf field: bool closed = 1;
     */
    closed: boolean;
    /**
     * 智能防挡弹幕蒙版信息
     *
     * @generated from protobuf field: bilibili.community.service.dm.v1.VideoMask mask = 2;
     */
    mask?: VideoMask;
    /**
     * 视频字幕
     *
     * @generated from protobuf field: bilibili.community.service.dm.v1.VideoSubtitle subtitle = 3;
     */
    subtitle?: VideoSubtitle;
    /**
     * 高级弹幕专包url(bfs)
     *
     * @generated from protobuf field: repeated string special_dms = 4;
     */
    specialDms: string[];
    /**
     * 云屏蔽配置信息
     *
     * @generated from protobuf field: bilibili.community.service.dm.v1.DanmakuFlagConfig ai_flag = 5;
     */
    aiFlag?: DanmakuFlagConfig;
    /**
     * 弹幕配置信息
     *
     * @generated from protobuf field: bilibili.community.service.dm.v1.DanmuPlayerViewConfig player_config = 6;
     */
    playerConfig?: DanmuPlayerViewConfig;
    /**
     * 弹幕发送框样式
     *
     * @generated from protobuf field: int32 send_box_style = 7;
     */
    sendBoxStyle: number;
    /**
     * 是否允许
     *
     * @generated from protobuf field: bool allow = 8;
     */
    allow: boolean;
    /**
     * check box 是否展示
     *
     * @generated from protobuf field: string check_box = 9;
     */
    checkBox: string;
    /**
     * check box 展示文本
     *
     * @generated from protobuf field: string check_box_show_msg = 10;
     */
    checkBoxShowMsg: string;
    /**
     * 展示文案
     *
     * @generated from protobuf field: string text_placeholder = 11;
     */
    textPlaceholder: string;
    /**
     * 弹幕输入框文案
     *
     * @generated from protobuf field: string input_placeholder = 12;
     */
    inputPlaceholder: string;
    /**
     * 用户举报弹幕 cid维度屏蔽的正则规则
     *
     * @generated from protobuf field: repeated string report_filter_content = 13;
     */
    reportFilterContent: string[];
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.ExpoReport expo_report = 14;
     */
    expoReport?: ExpoReport;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.BuzzwordConfig buzzword_config = 15;
     */
    buzzwordConfig?: BuzzwordConfig;
    /**
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.Expressions expressions = 16;
     */
    expressions: Expressions[];
    /**
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.PostPanel post_panel = 17;
     */
    postPanel: PostPanel[];
    /**
     * @generated from protobuf field: repeated string activity_meta = 18;
     */
    activityMeta: string[];
    /**
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.PostPanelV2 post_panel2 = 19;
     */
    postPanel2: PostPanelV2[];
}
// ...
export interface ExpoReport {
    /**
     * @generated from protobuf field: bool should_report_at_end = 1;
     */
    shouldReportAtEnd: boolean;
}
// ...
export interface Expressions {
    /**
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.Expression data = 1;
     */
    data: Expression[];
}
// ...
export interface PostPanel {
    /**
     * @generated from protobuf field: int64 start = 1;
     */
    start: bigint;
    /**
     * @generated from protobuf field: int64 end = 2;
     */
    end: bigint;
    /**
     * @generated from protobuf field: int64 priority = 3;
     */
    priority: bigint;
    /**
     * @generated from protobuf field: int64 biz_id = 4;
     */
    bizId: bigint;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.PostPanelBizType biz_type = 5;
     */
    bizType: PostPanelBizType;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.ClickButton click_button = 6;
     */
    clickButton?: ClickButton;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.TextInput text_input = 7;
     */
    textInput?: TextInput;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.CheckBox check_box = 8;
     */
    checkBox?: CheckBox;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.Toast toast = 9;
     */
    toast?: Toast;
}
/**
 * @generated from protobuf message bilibili.community.service.dm.v1.PostPanelV2
 */
export interface PostPanelV2 {
    /**
     * @generated from protobuf field: int64 start = 1;
     */
    start: bigint;
    /**
     * @generated from protobuf field: int64 end = 2;
     */
    end: bigint;
    /**
     * @generated from protobuf field: int32 biz_type = 3;
     */
    bizType: number;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.ClickButtonV2 click_button = 4;
     */
    clickButton?: ClickButtonV2;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.TextInputV2 text_input = 5;
     */
    textInput?: TextInputV2;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.CheckBoxV2 check_box = 6;
     */
    checkBox?: CheckBoxV2;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.ToastV2 toast = 7;
     */
    toast?: ToastV2;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.BubbleV2 bubble = 8;
     */
    bubble?: BubbleV2;
    /**
     * @generated from protobuf field: bilibili.community.service.dm.v1.LabelV2 label = 9;
     */
    label?: LabelV2;
    /**
     * @generated from protobuf field: int32 post_status = 10;
     */
    postStatus: number;
}
// ...
export interface VideoMask {
    /**
     * 视频cid
     *
     * @generated from protobuf field: int64 cid = 1;
     */
    cid: bigint;
    /**
     * 平台
     * 0:web端 1:客户端
     *
     * @generated from protobuf field: int32 plat = 2;
     */
    plat: number;
    /**
     * 帧率
     *
     * @generated from protobuf field: int32 fps = 3;
     */
    fps: number;
    /**
     * 间隔时间
     *
     * @generated from protobuf field: int64 time = 4;
     */
    time: bigint;
    /**
     * 蒙版url
     *
     * @generated from protobuf field: string mask_url = 5;
     */
    maskUrl: string;
}
/**
 * 视频字幕信息
 *
 * @generated from protobuf message bilibili.community.service.dm.v1.VideoSubtitle
 */
export interface VideoSubtitle {
    /**
     * 视频原语言代码
     *
     * @generated from protobuf field: string lan = 1;
     */
    lan: string;
    /**
     * 视频原语言
     *
     * @generated from protobuf field: string lanDoc = 2;
     */
    lanDoc: string;
    /**
     * 视频字幕列表
     *
     * @generated from protobuf field: repeated bilibili.community.service.dm.v1.SubtitleItem subtitles = 3;
     */
    subtitles: SubtitleItem[];
}
// ...
export const BuzzwordConfig = new BuzzwordConfig$Type();
// ...
export const DanmakuFlagConfig = new DanmakuFlagConfig$Type();
// ...
export const DanmuPlayerViewConfig = new DanmuPlayerViewConfig$Type();
// ...
export const DmViewReply = new DmViewReply$Type();
// ...
export const ExpoReport = new ExpoReport$Type();
// ...
export const Expressions = new Expressions$Type();
// ...
export const PostPanel = new PostPanel$Type();
// ...
export const PostPanelV2 = new PostPanelV2$Type();
// ...
export const SubtitleItem = new SubtitleItem$Type();
// ...
class VideoMask$Type extends MessageType<VideoMask> {
    constructor() {
        super("bilibili.community.service.dm.v1.VideoMask", [
            { no: 1, name: "cid", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 2, name: "plat", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "fps", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "time", kind: "scalar", T: 3 /*ScalarType.INT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 5, name: "mask_url", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<VideoMask>): VideoMask {
        const message = { cid: 0n, plat: 0, fps: 0, time: 0n, maskUrl: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<VideoMask>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: VideoMask): VideoMask {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int64 cid */ 1:
                    message.cid = reader.int64().toBigInt();
                    break;
                case /* int32 plat */ 2:
                    message.plat = reader.int32();
                    break;
                case /* int32 fps */ 3:
                    message.fps = reader.int32();
                    break;
                case /* int64 time */ 4:
                    message.time = reader.int64().toBigInt();
                    break;
                case /* string mask_url */ 5:
                    message.maskUrl = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: VideoMask, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int64 cid = 1; */
        if (message.cid !== 0n)
            writer.tag(1, WireType.Varint).int64(message.cid);
        /* int32 plat = 2; */
        if (message.plat !== 0)
            writer.tag(2, WireType.Varint).int32(message.plat);
        /* int32 fps = 3; */
        if (message.fps !== 0)
            writer.tag(3, WireType.Varint).int32(message.fps);
        /* int64 time = 4; */
        if (message.time !== 0n)
            writer.tag(4, WireType.Varint).int64(message.time);
        /* string mask_url = 5; */
        if (message.maskUrl !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.maskUrl);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message bilibili.community.service.dm.v1.VideoMask
 */
export const VideoMask = new VideoMask$Type();
// ...
export const VideoSubtitle = new VideoSubtitle$Type();