// src/protos/bilibili.pgc.gateway.player.v2/PlayURL.ts
export interface PlayViewReq {
    /**
     * 剧集epid
     *
     * @generated from protobuf field: int64 ep_id = 1;
     */
    epId: number;
    /**
     * 视频cid
     *
     * @generated from protobuf field: int64 cid = 2;
     */
    cid: number;
    /**
     * 清晰度
     *
     * @generated from protobuf field: int64 qn = 3;
     */
    qn: number;
    /**
     * 视频流版本
     *
     * @generated from protobuf field: int32 fnver = 4;
     */
    fnver: number;
    /**
     * 视频流格式
     *
     * @generated from protobuf field: int32 fnval = 5;
     */
    fnval: number;
    /**
     * 下载模式
     * 0:播放 1:flv下载 2:dash下载
     *
     * @generated from protobuf field: uint32 download = 6;
     */
    download: number;
    /**
     * 流url强制是用域名
     * 0:允许使用ip 1:使用http 2:使用https
     *
     * @generated from protobuf field: int32 force_host = 7;
     */
    forceHost: number;
    /**
     * 是否4K
     *
     * @generated from protobuf field: bool fourk = 8;
     */
    fourk: boolean;
    /**
     * 当前页spm
     *
     * @generated from protobuf field: string spmid = 9;
     */
    spmid: string;
    /**
     * 上一页spm
     *
     * @generated from protobuf field: string from_spmid = 10;
     */
    fromSpmid: string;
    /**
     * 青少年模式
     *
     * @generated from protobuf field: int32 teenagers_mode = 11;
     */
    teenagersMode: number;
    /**
     * 视频编码
     *
     * @generated from protobuf field: bilibili.pgc.gateway.player.v2.CodeType prefer_codec_type = 12;
     */
    preferCodecType: CodeType;
    /**
     * 是否强制请求预览视频
     *
     * @generated from protobuf field: bool is_preview = 13;
     */
    isPreview: boolean;
    /**
     * 一起看房间id
     *
     * @generated from protobuf field: int64 room_id = 14;
     */
    roomId: number;
    /**
     * 是否需要展示信息
     *
     * @generated from protobuf field: bool is_need_view_info = 15;
     */
    isNeedViewInfo: boolean;
    /**
     * 场景控制
     *
     * @generated from protobuf field: bilibili.pgc.gateway.player.v2.SceneControl scene_control = 16;
     */
    sceneControl?: SceneControl;
    /**
     * @generated from protobuf field: bilibili.pgc.gateway.player.v2.InlineScene inline_scene = 17;
     */
    inlineScene: InlineScene;
    /**
     * @generated from protobuf field: int64 material_no = 18;
     */
    materialNo: number;
    /**
     * @generated from protobuf field: int32 security_level = 19;
     */
    securityLevel: number;
    /**
     * @generated from protobuf field: int64 season_id = 20;
     */
    seasonId: number;
    /**
     * @generated from protobuf field: bilibili.pgc.gateway.player.v2.DataControl data_control = 21;
     */
    dataControl?: DataControl;
}
// ...
export interface SceneControl {
    /**
     * 是否收藏播单
     *
     * @generated from protobuf field: bool fav_playlist = 1;
     */
    favPlaylist: boolean;
    /**
     * 是否小窗
     *
     * @generated from protobuf field: bool small_window = 2;
     */
    smallWindow: boolean;
    /**
     * 是否画中画
     *
     * @generated from protobuf field: bool pip = 3;
     */
    pip: boolean;
    /**
     * @generated from protobuf field: bool was_he_inline = 4;
     */
    wasHeInline: boolean;
    /**
     * @generated from protobuf field: bool is_need_trial = 5;
     */
    isNeedTrial: boolean;
}
/**
 * @generated from protobuf message bilibili.pgc.gateway.player.v2.DataControl
 */
export interface DataControl {
    /**
     * @generated from protobuf field: bool need_watch_progress = 1;
     */
    needWatchProgress: boolean;
}
// ...
export enum CodeType {
    /**
     * 默认
     *
     * @generated from protobuf enum value: NOCODE = 0;
     */
    NOCODE = 0,
    /**
     * H.264
     *
     * @generated from protobuf enum value: CODE264 = 1;
     */
    CODE264 = 1,
    /**
     * H.265
     *
     * @generated from protobuf enum value: CODE265 = 2;
     */
    CODE265 = 2
}
/**
 * DRM技术类型
 *
 * @generated from protobuf enum bilibili.pgc.gateway.player.v2.DrmTechType
 */
export enum DrmTechType {
    /**
     * @generated from protobuf enum value: NON = 0;
     */
    NON = 0,
    /**
     * @generated from protobuf enum value: FAIR_PLAY = 1;
     */
    FAIR_PLAY = 1
}
/**
 * @generated from protobuf enum bilibili.pgc.gateway.player.v2.InlineScene
 */
export enum InlineScene {
    /**
     * @generated from protobuf enum value: UNKNOWN = 0;
     */
    UNKNOWN = 0,
    /**
     * @generated from protobuf enum value: RELATED_EP = 1;
     */
    RELATED_EP = 1,
    /**
     * @generated from protobuf enum value: HE = 2;
     */
    HE = 2,
    /**
     * @generated from protobuf enum value: SKIP = 3;
     */
    SKIP = 3
}
// ...
export const PlayViewReq = new PlayViewReq$Type();
// ...
class SceneControl$Type extends MessageType<SceneControl> {
    constructor() {
        super("bilibili.pgc.gateway.player.v2.SceneControl", [
            { no: 1, name: "fav_playlist", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 2, name: "small_window", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 3, name: "pip", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 4, name: "was_he_inline", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 5, name: "is_need_trial", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<SceneControl>): SceneControl {
        const message = { favPlaylist: false, smallWindow: false, pip: false, wasHeInline: false, isNeedTrial: false };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<SceneControl>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SceneControl): SceneControl {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* bool fav_playlist */ 1:
                    message.favPlaylist = reader.bool();
                    break;
                case /* bool small_window */ 2:
                    message.smallWindow = reader.bool();
                    break;
                case /* bool pip */ 3:
                    message.pip = reader.bool();
                    break;
                case /* bool was_he_inline */ 4:
                    message.wasHeInline = reader.bool();
                    break;
                case /* bool is_need_trial */ 5:
                    message.isNeedTrial = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SceneControl, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* bool fav_playlist = 1; */
        if (message.favPlaylist !== false)
            writer.tag(1, WireType.Varint).bool(message.favPlaylist);
        /* bool small_window = 2; */
        if (message.smallWindow !== false)
            writer.tag(2, WireType.Varint).bool(message.smallWindow);
        /* bool pip = 3; */
        if (message.pip !== false)
            writer.tag(3, WireType.Varint).bool(message.pip);
        /* bool was_he_inline = 4; */
        if (message.wasHeInline !== false)
            writer.tag(4, WireType.Varint).bool(message.wasHeInline);
        /* bool is_need_trial = 5; */
        if (message.isNeedTrial !== false)
            writer.tag(5, WireType.Varint).bool(message.isNeedTrial);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message bilibili.pgc.gateway.player.v2.SceneControl
 */
export const SceneControl = new SceneControl$Type();
// @generated message type with reflection information, may provide speed optimized methods
class DataControl$Type extends MessageType<DataControl> {
    constructor() {
        super("bilibili.pgc.gateway.player.v2.DataControl", [
            { no: 1, name: "need_watch_progress", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<DataControl>): DataControl {
        const message = { needWatchProgress: false };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<DataControl>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: DataControl): DataControl {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* bool need_watch_progress */ 1:
                    message.needWatchProgress = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: DataControl, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* bool need_watch_progress = 1; */
        if (message.needWatchProgress !== false)
            writer.tag(1, WireType.Varint).bool(message.needWatchProgress);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message bilibili.pgc.gateway.player.v2.DataControl
 */
export const DataControl = new DataControl$Type();